// src/app/models/types/user-profile.ts
export interface UserProfile {
  Profession: string;
  Country: string;
  State: string;
  PostalCode: string;
  City: string;
  Street: string;
  PortfolioPlatform: string;
  LinkedinPlatform: string;
  UpdatedBy: string;
  Guarantor1_FullName: string;
  GuarantorI_RelationshipToKin: string;
  GuarantorI_Phone: string;
  GuarantorI_EmailAddress: string;
  Guarantor2_Fullname: string;
  Guarantor2_RelationshipToKin: string;
  Guarantor2_PhoneNumber: string;
  Guarantor2_EmailAddress: string;
  AddressDocumentUrl: string;
  IsAddressVerified: boolean;
  IsGuarantorsVerified: boolean;
  CV_URL: string;
  AcademicAndProfessionalCertificateUrls: string;
  isPoofOfExpertiseVerified: boolean;
  IdDocumentUrl: string;
  IdDocumentType: string;
  isGovernmentIdDocumentVerified: boolean;
  SelfietUrl: string;
  isSelfieVerified: boolean;
  YearsOfExperience?:string;
  SalaryExpectationPerHour?:string;
}